feat: add comprehensive tests for SQL and TMQ failover scenarios
